Create wrapper hybrid
Graph observations, continuous actions
